# api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Agent Builder API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      responses:
        "200":
          description: OK
  /v1/agents:
    get:
      responses:
        "200":
          description: List agents
    post:
      parameters:
        - in: header
          name: X-Actor
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                owners:
                  type: array
                  items: { type: string }
                labels:
                  type: object
                  additionalProperties: { type: string }
      responses:
        "201":
          description: Created
  /v1/agents/{name}:
    get:
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200": { description: Agent }
  /v1/agents/{name}/versions:
    get:
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        "200": { description: Versions }
    post:
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: header
          name: X-Actor
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              type: string
      responses:
        "201": { description: Version created }
  /v1/versions/{id}:publish:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: header
          name: X-Actor
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      responses:
        "200": { description: Published }
  /v1/agents/{name}/runs:
    get:
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200": { description: Runs list }
    post:
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: header
          name: X-Actor
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputs]
              properties:
                inputs:
                  type: object
      responses:
        "202": { description: Run started }